move from doubles to floats

+++ glmc4-client/src/main/java/p0nki/glmc4/client/assets/AtlasPosition.java
@@ -2,16 +2,16 @@ package p0nki.glmc4.client.assets;
 
 public class AtlasPosition {
 
-    public double x;
-    public double y;
-    public double w;
-    public double h;
+    public float x;
+    public float y;
+    public float w;
+    public float h;
 
     public AtlasPosition(int x, int y, int w, int h, int tw, int th) {
-        this.x = (double) x / tw;
-        this.y = (double) y / th;
-        this.w = (double) w / tw;
-        this.h = (double) h / th;
+        this.x = (float) x / tw;
+        this.y = (float) y / th;
+        this.w = (float) w / tw;
+        this.h = (float) h / th;
     }
 
 }
